Rename ChatGPT folder to OpenAI. Basic alert handling WIP.
