+++ Engage/ViewModels/ChatViewModel.cs
@@ -2,9 +2,8 @@
 using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
-using Engage.ChatGPT;
+using Engage.OpenAI;
 using Engage.Models;
-using Microsoft.Extensions.Logging;
 
 namespace Engage.ViewModels
 {
