+++ Engage/Views/ChatPage.xaml
@@ -3,12 +3,10 @@
     x:Class="Engage.Views.ChatPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:controls="using:Engage.Views.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:viewModels="using:Engage.ViewModels"
-    xmlns:vm="using:Engage.ViewModels"
     mc:Ignorable="d">
 
     <Grid>
@@ -75,52 +73,62 @@
             </HyperlinkButton>
         </Grid>
 
-        <!--  Chat Area  -->
-        <ScrollViewer
-            x:Name="ChatScrollViewer"
-            Grid.Row="1"
-            Padding="10"
-            HorizontalScrollBarVisibility="Disabled"
-            VerticalScrollBarVisibility="Auto">
-            <ItemsControl x:Name="ChatItemsControl" ItemsSource="{Binding ElementName=ChatTabView, Path=SelectedItem.Messages}">
-                <ItemsControl.ItemTemplate>
-                    <DataTemplate>
-                        <Grid Margin="10">
-                            <Grid.RowDefinitions>
-                                <RowDefinition Height="Auto" />
-                                <RowDefinition Height="*" />
-                            </Grid.RowDefinitions>
+        <!--  Chat Area and Alert  -->
+        <Grid Grid.Row="1">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="*" />
+            </Grid.RowDefinitions>
 
-                            <TextBlock
-                                Grid.Row="0"
-                                Margin="0,0,0,10"
-                                HorizontalAlignment="{Binding ChatMessageType, Converter={StaticResource RoleToAlignmentConverter}}"
-                                VerticalAlignment="Top"
-                                FontSize="12"
-                                FontWeight="SemiBold"
-                                Text="{Binding Role, Converter={StaticResource StringToTitleCaseConverter}}" />
+            <controls:Alert Grid.Row="0" />
 
-                            <Border
-                                x:Name="ChatBorder"
-                                Grid.Row="1"
-                                MaxWidth="700"
-                                Padding="10"
-                                HorizontalAlignment="{Binding ChatMessageType, Converter={StaticResource RoleToAlignmentConverter}}"
-                                Background="{Binding ChatMessageType, Converter={StaticResource ChatMessageTypeToBackgroundColorConverter}}"
-                                BorderThickness="1"
-                                CornerRadius="5">
+            <!--  Chat Area  -->
+            <ScrollViewer
+                x:Name="ChatScrollViewer"
+                Grid.Row="1"
+                Padding="10"
+                HorizontalScrollBarVisibility="Disabled"
+                VerticalScrollBarVisibility="Auto">
+                <ItemsControl x:Name="ChatItemsControl" ItemsSource="{Binding ElementName=ChatTabView, Path=SelectedItem.Messages}">
+                    <ItemsControl.ItemTemplate>
+                        <DataTemplate>
+                            <Grid Margin="10">
+                                <Grid.RowDefinitions>
+                                    <RowDefinition Height="Auto" />
+                                    <RowDefinition Height="*" />
+                                </Grid.RowDefinitions>
+
                                 <TextBlock
-                                    FontSize="14"
-                                    Foreground="{Binding ChatMessageType, Converter={StaticResource ChatMessageTypeToForegroundColorConverter}}"
-                                    IsTextSelectionEnabled="True"
-                                    Text="{Binding Content}"
-                                    TextWrapping="Wrap" />
-                            </Border>
-                        </Grid>
-                    </DataTemplate>
-                </ItemsControl.ItemTemplate>
-            </ItemsControl>
-        </ScrollViewer>
+                                    Grid.Row="0"
+                                    Margin="0,0,0,10"
+                                    HorizontalAlignment="{Binding ChatMessageType, Converter={StaticResource RoleToAlignmentConverter}}"
+                                    VerticalAlignment="Top"
+                                    FontSize="12"
+                                    FontWeight="SemiBold"
+                                    Text="{Binding Role, Converter={StaticResource StringToTitleCaseConverter}}" />
+
+                                <Border
+                                    x:Name="ChatBorder"
+                                    Grid.Row="1"
+                                    MaxWidth="700"
+                                    Padding="10"
+                                    HorizontalAlignment="{Binding ChatMessageType, Converter={StaticResource RoleToAlignmentConverter}}"
+                                    Background="{Binding ChatMessageType, Converter={StaticResource ChatMessageTypeToBackgroundColorConverter}}"
+                                    BorderThickness="1"
+                                    CornerRadius="5">
+                                    <TextBlock
+                                        FontSize="14"
+                                        Foreground="{Binding ChatMessageType, Converter={StaticResource ChatMessageTypeToForegroundColorConverter}}"
+                                        IsTextSelectionEnabled="True"
+                                        Text="{Binding Content}"
+                                        TextWrapping="Wrap" />
+                                </Border>
+                            </Grid>
+                        </DataTemplate>
+                    </ItemsControl.ItemTemplate>
+                </ItemsControl>
+            </ScrollViewer>
+        </Grid>
 
 
         <!--  CONTROLS  -->
