+++ Engage/Views/Controls/Alert.xaml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<UserControl
+    x:Class="Engage.Views.Controls.Alert"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="using:Engage.Views.Controls"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <!--  Alert control  -->
+    <StackPanel
+        Width="Auto"
+        HorizontalAlignment="Center"
+        VerticalAlignment="Top"
+        Orientation="Horizontal">
+        <Border
+            Margin="10"
+            Padding="10"
+            Background="{StaticResource SystemFillColorCautionBackgroundBrush}"
+            CornerRadius="4">
+            <Grid ColumnSpacing="10">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*" />
+                    <ColumnDefinition Width="Auto" />
+                </Grid.ColumnDefinitions>
+                <TextBlock
+                    Grid.Column="0"
+                    HorizontalAlignment="Center"
+                    VerticalAlignment="Center"
+                    Text="You have exceeded your monthly spending limit." />
+                <Button
+                    x:Name="CloseAlertButton"
+                    Grid.Column="1"
+                    Width="Auto"
+                    HorizontalAlignment="Center"
+                    VerticalAlignment="Center"
+                    Click="CloseAlertButton_Click">
+                    <FontIcon
+                        FontFamily="Segoe MDL2 Assets"
+                        FontSize="9"
+                        Glyph="&#xEF2C;" />
+                </Button>
+            </Grid>
+        </Border>
+    </StackPanel>
+</UserControl>
