+++ Engage/Views/MainWindow.xaml.cs
@@ -1,7 +1,7 @@
 // Engage.Views.MainWindow.xaml.cs
 using System;
 using System.Diagnostics;
-using Engage.ChatGPT;
+using Engage.OpenAI;
 using Engage.Helpers;
 using Engage.ViewModels;
 using Microsoft.UI.Xaml;
